Regression: Some QMenus are shown completely transparent in Symbian
Fix for menu transparency (QTBUG-16857) unveiled an issue in the
QS60Style when drawing QMenu background.

Style skips drawing of the menu background, since placeHolder
texture for QPalette::Window is not replaced with real background
for menus. Therefore, we need to check if the widget's palette
is still untouched when deciding whether or not to draw the
menu background.

Task-number: QTBUG-20255
Reviewed-by: Tomi Vihria