Avoid crash when surface creation fails.
When the creation of the EGL window surface failed, the paint engine
was not destroyed in the openvg module's window surface implementation.
This led to accessing an obsolete engine instance later on in
QSymbianControl::Draw, leading to a crash. Now a new, proper engine
will be created, naturally the painting will still fail, but the
application will not crash.

Task-number: QTBUG-20339
Reviewed-by: Jani Hautakangas