Get subpixel antialiasing again w/combo of raster and affine transform
Revert part of hot fix 48ff7e5af99923396243940979d15d4ec2e2730c since
it is no longer needed in Qt 4.8 due to the introduction of a separate
code path for the FreeType case in drawCachedGlyphs(). This is also
explained in the description for that commit. The consequence of the
commit was that subpixel antialiasing would be turned off for all
text drawn with a transformation and the raster engine. In the past
this was not the case, it is however the case in several other paint
engine, so the regression was deemed okay to hot fix the original
problem. In Qt 4.8 the original problem has been fixed properly.

Task-number: QTBUG-16205
Reviewed-by: Jiang Jiang