Save previous font engine for right bearing adjustment
In last fix I forgot that fd818312 was for saving and restoring the
right bearing (of last visible glyph) when a LineSeparator was hit
(which can have a different font engine but usually not visble),
thus we can't reset previousGlyph in that case. To make sure we
still get correct right bearing from the font engine used to shape
previousGlyph, we need to save that font engine as well. It does
make the code more complicated than simply saving the right bearing
when a QScriptItem boundary is hit, so hopefully it's an optimization
worth to be made (following e1915815).

Task-number: QTBUG-20423
Reviewed-by: Eskil