Detect linked fonts by name (insead of via CLinkedTypeface* Api)
The detection of Linked Fonts via CLinkedTypeface* Api does not
work in all cases. It works in QML apps, but not in QWidgets based
apps. The reason is still unclear.

This commit replaces the CLinkedTypeface* Api based code of commit
fa344b355d311fce7954e2fd1a22a87a88194783 whith a simple type face
anme analysis. Assumption is that the names of Linked Fonts always
end with "LF" and are all upper case.

This patch has been tested by more people than just me, but it is
too dirty to mention the others as reviewer. I take the shame and
blame.

Task-Number: QTBUG-20007