Preventing QSoftkeyManager giving false positive memory leaks
QSoftkeyManager has a global static instance that was not being
deleted on app exit. This global static instance can own other objects,
the number of which grow with use up to a limit. This gives the
appearance of a memory leak when the app exits and you see increasing
heap cell counts.

The change is to use a QScopedPointer to clean up the static on app
exit.

Reviewed-by: Miikka Heikkinen