Automatic closing of input panel on uikit when QML item looses focus.
I.e. we remember which item had focus when we got the request to open
the input panel, and close it when that looses focus, even if we don't
get an explicit close request.

Also set the "returnKey" to "Return" instead of "Done" if the focus item
is a TextEdit.

This adds a dependency of the platform plugin to declarative.