Re-introduce Q_GLOBAL_STATIC_INIT
Removing this macro in abff4d5090c1706c44485bbe3689a0e339c26a9b breaks
code that used it in providing their own class-scope safe
Q_GLOBAL_STATIC variations, the feature introduced in that commit. There
may be other use cases for the macro.

Anyway, even if this isn't part of the public API, there isn't a strong
reason to remove the macro, so it is back now.

Task-number: QTBUG-20627
Reviewed-by: Samuel Rødal