Cocoa: change usage of deprecated API (scrolling) in Lion
Apple changed the API for pixel based scrolling, which results
in runtime warnings on the command line. This patch will make
Qt use the new API. In addition, apple removed suppert for
z-scrolling. Since that was never supported in Qt as well, we
also remote code that handles that.

RevBy: msorvig