tests: fixed crash of tst_qdialog
This test assumed that C++ exceptions could always be caught by the
event loop.  This is not the case when the Glib event loop is used.
Skip the relevant portion of the test in that case.

Change-Id: I2ad83386025a98f6c32202700c97e04ef3dff343
Reviewed-by: Kalle Lehtonen
Reviewed-by: Sergio Ahumada

(cherry picked from commit 569cd194d20e61d35768b210b4351f4eeb5c1ef8)