Prevent leakage of native window handles
On Symbian, reparenting a native widget causes its native window,
and those of all its ancestors, to be destroyed and recreated.

Because native window handles cannot be destroyed from the context
of CONE event handlers, the destruction is delayed until after
control flow returns to the event loop (see QTBUG-4664).

However, the pending-destruction native windows are leaked if
either of the following happens:

a) The application never returns to the event loop.  While unlikely
in a real app, this situation does happen in Qt autotests, causing
a crash in tst_qwidget due to leakage from
tst_QWidget::reparentCausesChildWinIdChange().

b) A subsequent reparenting event occurs before control returns to
the event loop.

This patch prevents the leak by storing the pending-destruction
native window handle(s) in a member QList, rather than on the stack
as parameters to QWidgetPrivate::_q_delayedDestroy(WId).

Task-number: QT-5135
Reviewed-by: Jani Hautakangas