Symbian: Fix backspace on empty lines of multiline textedits
Backspace on empty lines of multiline textedits didn't work because
in Symbian the virtual keyboard only gets one block of text at a time,
and there is some internal optimization in FEP that if the document
is empty, backspace doesn't get generated. Fixed by faking document
lenght to be one in cases where it is actually zero, except for password
fields with hidden text, as an extra asterisk would be generated in
those cases. Typically password fields are not multiline so this
shouldn't be a problem.

Task-number: QTBUG-20444
Reviewed-by: Sami Merila