Do not unconditionally use pkg-config in .pro files
Using the first pkg-config in PATH is prone to errors when
cross compiling.  This is a known issue, and `configure' already
avoids using pkg-config for cross compiling (unless -force-pkg-config is
used).  However, there was previously no way for a .pro file to test
whether or not the user had passed -force-pkg-config.

Let !contains(QT_CONFIG, no-pkg-config) be used to test whether or not it
is safe to run pkg-config.