Avoid calling QInputMethod update when adding or removing QGraphicsItems that don't have focus
Removing item from QGraphicsScene takes longer if focus is in item (and that item is not being removed) with ItemAcceptsInputMethod.
update() in QInputContext or in subclasses could be expensive - usually requires inputMethodFocus and a sequence of inputMethodQuery calls to get the state. In MeeGo, additionally, dbus IPC in MInputContext makes it order of magnitude slower.
Minor indentation issue in QGraphicsViewPrivate::updateInputMethodSensitivity()

Reviewed-by: Samuel Rødal