Make it easier to select words at the start of a line.
QTextControl's word selection will only include a word if the cursor
position is past the mid point of the word. This can make it
difficult to select words near the edges of the screen on touch
devices.  For the TextEdit word selection mode select a word ignore
the relative position within a word.

Change-Id: I2bc61376f663836fedd7e157448f0b565a64d485
Task-number: QT-5206
Reviewed-by: Martin Jones