URLs used with scheme handler use appropriate encoding for scheme
URLs passed to scheme handler always used to use raw encoding.
That broke URLs containing spaces passed to the browser.

Then QTBUG-18772 changed it so that they always used % encoding.
That broke URLs containing spaces used to open files, resulting
in QTBUG-21209.

This change fixes it so that web URLs are % encoded and file
URLs are raw encoded.

Task-number: QTBUG-21209
Reviewed-by: Honglei Zhang