Ensure that the proxystyle is used over the style's own functions
When calling a style function such as pixelMetric() then this should be
done via the proxy style and not on the style directly.  This is so that
the proxy style always has a chance to override the original style's
implementation and still preserve the fallback to the base style.

Task-number: QTBUG-20849
Merge-request: 2645
Reviewed-by: olivier