Fix RTL layout for fonts when non-printable character has an advance
If a non-printable character has an advance, we would lay out the text
as if the character was part of the output, and then ignore the
character later when the text was rendered. To fix cases such as this,
we take the dontPrint attribute into consideration when adjusting the
advances.

Task-number: QTBUG-21347
Reviewed-by: Eskil
Reviewed-by: Lars