StrictlyEnforceRange with snapOneItem/Row and header behavior change
Change cf23188de237009136fa1480ab8fd9e3ca364769 changed the positioning
of a view with StrictlyEnforceRange, snapOneItem/Row, and a header,
such that the view was positioned at the beginning of the header,
rather than on the first item.  This change reverts back to the old
behavior (position on the first item).

Change-Id: I62ad183919bb2ed83d787d1d76421caf9e708599
Task-number: QTTH-1501
Reviewed-by: Michael Brasser