Fix uninitialised variable in temporary files
The symbianFilePos variable was only initialised in the nativeOpen
function and not in the constructor called init(). Due to the fix
for QTBUG-4796, QTemporaryFile now bypasses the nativeOpen function.
Zero initialising in init() ensures it is always valid (it only needs
a non zero initialisation when opening a file for append, which is
still covered by nativeOpen)

Task-Number: QTBUG-21418
Reviewed-By: mread