Fix StrictlyEnforceRange with snapOneItem/Row and header behavior, pt 2
Change cf23188de237009136fa1480ab8fd9e3ca364769 changed the positioning
of a view with StrictlyEnforceRange, snapOneItem/Row, and a header,
such that the view was positioned at the beginning of the header,
rather than on the first item.

Change f85819fe083ae7c6804c884de68e906d153a6d11 partially fixed the
problem.  This change handles the case of the header/footer being
large enough to cause a snap item not to be found when the view
is dragged beyond the first/last item.  In this case snap to the
currentItem.

Change-Id: I08b7e61496a79f71c3b40fafaca985ae90f88503
Task-number: QTTH-1501
Reviewed-by: Bea Lam