Prevent crash when cache is changed on the fly
Calling QNetworkAccessManager::setCache while there were requests
being processed using the existing cache caused crashes due to
deleting the old cache invalidating pointers to the cache data
streams (QTemporaryFile in case of QNetworkDiskCache).
Using pointer to deleted data caused a crash in some cases. It has
undefined behaviour because the memory may have been overwritten or
decommitted.

To fix this, use the deleted signal to notify QNetworkReplyImpl that
the cache has been destroyed. It then clears the pointer to the data
stream and disables caching. This avoids the crash that previously
happened when trying to write to the cache file.

Task-number: QT-5252
Reviewed-by: Peter Hartmann