symbian bearer: fix tst_qnetworksession test failures
Symbian is not reporting the expected KLinkLayerClosed event on the
last RConnection handle. Instead, it reports the event
KConfigDaemonFinishedDeregistrationStop, which indicates the nif
plugin is being destroyed (this is followed by unload events)

To fix, this event is treated the same as a link down event, and
results in the state change to disconnected.

Either there has been a behaviour change in the platform, or the
test was previously run with another user of the API running in the
background (e.g. WLAN wizard, connection monitor) which prevented
the unload.

Task-Number: QTBUG-21382
Reviewed-By: mread