Symbian: Added copy-paste functionality to FEP input context
The support for copy-paste functionality in FEP was initially submitted
as merge request #1151, though that only provided support for
basic Qt edit widgets. I modified the code so that QML edit elements
are also supported, as well as any other editor that provides the
required "copy", "cut", and "paste" slots, and "canPaste" property.
A number of other minor fixes were also done to the original merge
request, as well as the removal of S60 version plugin elements.

Task-number: QTBUG-20921
Reviewed-by: Sami Merila