QML import path puts Qt dirs in correct Symbian order
The QML import path contains an entry for the standard
import location on each drive on Symbian. These were
being ordered so that Z:, the ROM drive, was listed
first. In fact the drive order was the reverse of the
normal Symbian drive search order. This had the effect
that upgraded versions of QML imports installed on
the device were generally ignored for older ROM based
versions.

The import path has now been changed to match the normal
Symbian drive search order. This will make it more likely
for upgraded QML imports to be picked up and used.

Task-number: QTBUG-21409
Reviewed-by: Shane Kearns
Reviewed-by: Martin Jones