QLineControl: selectedTextChanged() signal is omitted in some cases
QLineControl does not emit selectionChanged() signal when
selection is changed due to handling of input method query.

As a fix, the signal is emitted when internal values for selection
end and beginning are changed in processInputMethodEvent().

Additionally, added deselect() test case for both QDeclarativeTextEdit
and QDeclarativeTextInput to help out autotesting of text selections.

Note that there are some differences in the QLineControl and QTextControl
implementations. This causes the different signal counts even if the
test steps are similar.

Task-number: QTBUG-21429
Reviewed-by: Miikka Heikkinen