runonphone - Implement CODA autodetection
If runonphone is executed without specifying the debug agent, then
it will attempt to autodetect it by sending a CODA ping to the USB
port.
If there is a reply within 1 second, then CODA mode is used.
If not, then TRK mode is used.

TRK drops unrecognised messages, so the CODA ping is ignored and
initialisation starts normally when the TRK ping is sent.

Autodetect can be skipped by using the --coda or --trk arguments on
the command line to force use of a specific debug agent.

Reviewed-By: mread