The cursor can get offscreen in splitscreen landscape mode
When input context is making the translation, it doesn't check if
the translation is happening to the correct direction (i.e. upwards
or downwards).

As a fix, check if the cursor is still visible after making the
translation. If it isn't apply the translation to the other direction.

Task-number: QTBUG-21543
Reviewed-by: Miikka Heikkinen