Symbian - perform some FS initialisation previously done by open c
When using Qt 4.6 and 4.7, the working directory is set to the same
drive the application is installed on by the open c library.
The default working directory in symbian is always the private
directory on the system drive, but we told people to rely on the
open c behaviour.

Open C also creates the private path on both the system and
installation drives when the application is started.

This behaviour is also replicated now in Qt 4.8 for backward
compatibility of apps that may rely on it.

Similar code to create the private path on the installation drive
in QCoreApplication::applicationDirPath() has been removed, as the
new code in this patch is always executed first.

Task-Number: QTBUG-21527
Reviewed-By: mread