Making QNotifyChangeEvent, Symbian file watcher, more widely usable
QNotifyChangeEvent was being used by QSymbianFileSystemWatcherEngine,
but it also helps in the implementation of new plugin watching for
QTBUG-20098. So it has been generalised to work with an interface and
any class that implements that interface.

The Qt file system watchers could not be used in QTBUG-20098, as they
will not watch for currently non-existing directories.

Task-number: QTBUG-20098
Reviewed-by: Shane Kearns