Added Symbian private API to update libraryPaths
Library paths on Symbian can change as memory cards are inserted or
removed. Potentially there is a \resource\qt\plugins on a card which
should be added to the library path.

Since some Symbian Qt applications are expected to be very long running,
they would not see an updated library path without this abililty to
update the path.

The update algorithm strips out existing instances of
\resource\qt\plugins, places new ones where the first
\resource\qt\plugins was, and leaves the rest of the library path
unchanged.

Task-number: QTBUG-20098
Reviewed-by: Shane Kearns