Fixed broken window surface flush when depth is 24 and bpp is not 32.
Some X servers use a compact representation of 24 depth visuals. In that
case we can't use the shared memory or XPutImage paths, as Qt's RGB32
does not match the internal memory layout.

Also fixed QPixmap::fromImage() to work in this case to prevent the red
and blue channels from being swapped.

Task-number: QTBUG-21754
Reviewed-by: Alberto Mardegan