Do not crash in copy() of pixmaps without an underlying bitmap
If CFbsBitmap::Create() fails for some reason (e.g. due to lack of
memory), we may end up with a QVolatileImage for which the underlying
bitmap pointer is null, resulting in the QImage wrapper being null
too. The copyFrom() function was not checking for this situation and
started to copy data blindly to the null QImage's bits(), which is
a null pointer. This is now fixed so no copying occurs in such a
scenario.

Task-number: QTTH-1446
Reviewed-by: Sami Merila