QNAM - fix poor performance of HEAD request with authentication
QHttpNetworkReply was waiting for a body to be sent for 401 and 407
responses, whereas with a HTTP HEAD request, there will be no body.

This delayed the authentication signal until after the http channel
is closed by the server after a timeout. For example with the server
used for autotesting, the authentication signal is delayed 15 seconds.

When the server has a very long timeout, the authentication signal may
not be emitted at all.

Task-Number: QT-5304
Reviewed-By: Martin Petersson