Symbian: Fix native dialog with splitscreen VKB focus loss issue
If a native input dialog that utilized splitscreen VKB was opened while
focus was on QML TextEdit element, the focus would not be restored to
QML TextEdit element after native dialog was closed. This happened
because the KSplitViewCloseEvent came after FocusChanged, as the focus
changes are ignored while splitview VKB is open.

Task-number: QTBUG-21733
Reviewed-by: Sami Merila