Copy and Cut are not available for some QML editors
Copy and Cut actions do not appear in fullscreen VKB menu when a QML
TextEdit or TextInput element is focused and has selected text.
This happens because input method queries for cursor and anchor
position in QCoeFepInputContext::CcpuCanCut() return zero in this case.
It is probably related to focus changes when the options menu pops up.

When menu pops up from native keyboard, window containing the QML
elements is set as non-active, therefore it loses focus widget, which
in turn causes queries about cursor and anchor positions to return zero.

As a workaround, when there is no microfocus available, ask the
"selectedText" property directly from QML elements.

Task-number: QTBUG-21568
Reviewed-by: Miikka Heikkinen