Converting accidental use of doubles to qreal in declarative
Declarative had a number of places where double operations were used
instead of qreal, due to the use of double literals. These are now
constructed as qreal literals so that qreal operations are used.

This makes no difference where qreal is double. But on Symbian, qreal
is float, and this give a performance boost for floating point intensive
code.

Task-number: QTBUG-4894
Reviewed-by: Martin Jones