Removing accidental use of double instead of qreal from QLine
QLine contained a number of places where conversions and calculations
were done as double where qreal should have been used instead. This
is caused by the use of raw floating point literals. The fix is to
construct the literals as qreal. The angle conversions are simplified
to reduce the risk of a growing loss of precision from this change.

On Symbian, where qreal is float, this gives a 5-10% performance
improvement to the affected functions, as tested by a new benchmark
test. The auto tests pass, these contain precision tests which indicates
that there is no significant loss of precision with this change.

On Windows, where qreal is double, this has no significant effect on
performance.

Task-number: QTBUG-4894
Reviewed-by: Shane Kearns