Converting from double to qreal in gui
There were a number of places in QtGui where doubles were used in
expressions due to the use of floating point constants in the
code. Many of these constants are now constructed as qreal, removing
the use of double operations where unneeded.

These changes have been limited to constants that have exactly the same
value whether double or float, to ensure that precision errors are not
introduced.

This should not affect any of the desktop platforms where qreal is
double. On Symbian, where qreal is float, appropriate autotests have
been run.

Task-number: QTBUG-4894
Reviewed-by: Sami Merila