Symbian - fix deleteLater not working from RunL
deleteLater stores the loop level in the deferred delete event to
prevent the object being deleted by a nested event loop.
However as symbian active object RunL functions are called directly
from the active scheduler, the loop level is incorrect at that point.
(It is normally set by QCoreApplication::notifyInternal)

To solve this, the loop level is adjusted before calling RunIfReady
so that it is correct during RunL functions. It is then adjusted back
for the specific active objects in the event dispatcher that call
into QCoreApplication - sendPostedEvents, sendEvent.

Task-number: QTBUG-21928
Reviewed-by: mread