symbian - search drives for translation files
Qt may be installed on a different drive from the application,
particularly the case when Qt is included in ROM (Z:) and the
application is on C:

With this change, if QTranslator::load() specifies an absolute
directory in the filesystem (e.g. "/resource/qt/translations") without
a drive letter, then the symbian search paths are used.
Note that this example path is the one returned by QLibraryInfo so
applications using the example code from
http://doc.qt.nokia.com/latest/internationalization.html#produce-translations
will work as expected.

Task-number: QT-5246
Reviewed-by: mread