Accepting predicted text using hardware keyboard replaces unwanted part
Current implementation of Symbian input context assumes that predicted
word replacement happens so that the original typed text is at the end
of the surrounded text. The logic fails, if to-be-predicted text is
in the middle of, or in the beginning of another, already accepted
word.

As a fix, input context need to store the original cursor position,
when reset() was called (this happens when word selection list
appears). Input context is already storing a copy of a preedit string
in this situation.

Then, when word replacement happens, this stored cursor position is
used instead of current cursor position (the native side might
temporarily move the cursor to the end when word selection list opens
or closes) to replace the typed word with one selected from suggested
word list.

Stored cursor position is dismissed immediately after used, or
if cached preedit string is dismissed.

Task-number: QTBUG-22147
Reviewed-by: Miikka Heikkinen