Notify X11 of its resizable state through MWM Hints.
We only updated the MWM_FUNC_RESIZE when the window was shown.
The consequence was that if the widget was initially shown in fixed
size it would clear the MWM_FUNC_RESIZE flag, and we would never
set it again if the widget became resizable later.
Some window managers still support this (apparently it seems that
compiz tried to adhere to this)

Task-number: QTBUG-21914
Reviewed-by: Denis Dzyubenko <denis.dzyubenko@nokia.com>