Ensure that using multiple GridViews with the same model source works
When the same model source is used in multiple GridViews then it can
get into a state where items that have been inserted are removed from
a list which it then later tries to access to emit signals for.

This patch ensures that a list of the items that the signal is emitted
for is kept before any of the wrapper items are potentially deleted so
that the signals can be emitted as expected correctly.

Task-number: QTBUG-22271