Update supported platforms and remove commercial page.
The supported platforms page conatined a lot of platforms that are
untested for Qt 4.8, so those are removed. The platform notes had not
been updated since 2009 and were therefore severely outdated.

Nokia does not provide a commercial edition of Qt anymore. A user can
find the link to purchasing Qt through the Qt Developer Network at
http://developer.qt.nokia.com