Move tga support from Qt3d to Qt.
Rev-By: Samuel Rodal
Fix-for: QTBUG-21955

Part of fix for https://bugreports.qt.nokia.com/browse/QTBUG-21955

Having tga plugin inside a Qt add-on causes packaging problems.

There have been many queries over the years for tga support even tho it
is a niche image format.  It is particularly useful for OpenGL work as a
texture format since it is a blittable bitmap, but it is very
innefficient compared to png or jpg.

For this reason only read support is added, and this is documented.
Also add some unit tests.

The unit tests are in the qimagereader suite, and this is the right
place for them as the other formats are tested there.

Noticed that there was an annoying bug with tiff files and chased it
down, in the hope of getting a clean test case PASS.  Looks as tho
there is a bit of work for someone to fix the other FAILs, but the ones
for tga are all passing.