Increase qdoc3 stack size when using MSVC to 4M.
When running 'make docs' on Win7 with MSVC2010 64bit debug qdoc will
crash from a stack overflow. The stack overflow occurs when sorting the
list of nodes in the internal qdoc tree.

This patch circumvents the stack overflow crash by increasing the stack
size from the standard 1M to 4M.

Reviewed-by: Friedemann Kleint