Do not generate -D defines for project types.
Options are implicitly inherited from project to subprojects
and applications, so there is no need to respecify them.
Worse, if an app/lib disables an option, the project might
still have the define enabled, forcing it also onto the app/lib.
This prevents double-define or this kind of conflict.

Merge-request: 1438
Reviewed-by: Harald Fernengel <harald.fernengel@nokia.com>