Add XPLATFORM_INTEGRITY to define defaults for INTEGRITY.
Defaults are set for QT_INSTALL paths, and for various CFG_ defines.
Also set the input and framebuffer drivers to be used.
Also set the BUILD_PARTS to the reduced set that produce target code.
If using script and declarative, then these are added still.
Make the configuration explicitly "Qt for INTEGRITY".
Since *_DIR variables are set in the first part of the configuration, do not set them again in the generic code.

Merge-request: 1438
Reviewed-by: Harald Fernengel <harald.fernengel@nokia.com>