Export QtGui functions required by QtMultimediaKit backend
In order for video rendering to work, the Symbian QtMultimediaKit
backend needs access to QtGui functionality which is not exposed
via public interfaces, namely:
* QSymbianControl::setIgnoreFocusChanged(bool)
* Setting custom painting modes, which control how content is
  blitted from the backing store into the native window surface
* Requesting notification when Qt drawing starts and stops, so
  that the backend can ensure this does not interfere with video
  rendering.

Until now, this has been achieved by the backend including private
QtGui headers.  If these headers are not present in the environment
in which QtMobility is compiled, a warning is printed, telling the
user that video / viewfinder rendering will not work correctly when
the resulting binaries are installed.

This patch adds a Symbian-specific public interface to QtGui, which
exports the functionality above.

Task-number: MOBILITY-747
Reviewed-by: mread