Surviving out of memory in Qt Quick app
The QtQuickPlayground app contains a version of samegame which allow
the user to edit the code. By setting the ball size to 8, the app can
run out of memory. This leaves it in a pretty bad state. But apps
on Symbian shouldn't crash due to OOM and should allow some operation.

This change fixes the immediate OOM crashes in declarative, gui and
corelib. It shows warning dialogs which explain what has gone wrong and
leaves the app in a state that can be exited cleanly
from the Symbian task list.

Task-number: QT-5319
Reviewed-by: Shane Kearns
Reviewed-by: Gareth Stockwell
Reviewed-by: Martin Jones